move pr related utilities into the pr_list subpackage
